Fix signed integer truncation in Serializer::read
read(int16&), read(int32&) and read(int64&) cast the full-width value
to int8, truncating every serialized signed field to 8 bits (e.g. an
int32 of 100000 was read back as -96). The write side already stores
the full width, so correcting the cast is backward compatible with
existing cache files.

Export Serializer and MemorySerializer (IG_LIB) so the runtime's
serialization API is linkable from tests, and add a round-trip
regression test covering signed/unsigned integers and floats.

+++ src/runtime/serialization/Serializer.inl
@@ -199,7 +199,7 @@ inline void Serializer::read(int16& v)
 {
     uint16 tmp;
     read(tmp);
-    v = static_cast<int8>(tmp);
+    v = static_cast<int16>(tmp);
 }
 
 inline void Serializer::read(uint16& v)
@@ -211,7 +211,7 @@ inline void Serializer::read(int32& v)
 {
     uint32 tmp;
     read(tmp);
-    v = static_cast<int8>(tmp);
+    v = static_cast<int32>(tmp);
 }
 
 inline void Serializer::read(uint32& v)
@@ -223,7 +223,7 @@ inline void Serializer::read(int64& v)
 {
     uint64 tmp;
     read(tmp);
-    v = static_cast<int8>(tmp);
+    v = static_cast<int64>(tmp);
 }
 
 inline void Serializer::read(uint64& v)
